dashboard/requirements.txt
streamlit
psycopg2-binary
# any other libs you use

